requirements.txt
jsonschema==4.25.1
pytest==8.4.2
requests==2.32.5
allure-python-commons
selene==2.0.0rc9
allure-pytest==2.15.0